requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Telegram
python-telegram-bot[job-queue]==21.0.1

# Object Storage (S3/MinIO)
boto3==1.34.34
minio==7.2.3

# YouTube API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# HTTP Client
httpx>=0.27.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Security
python-jose[cryptography]==3.3.0

# Video processing (optional, для будущих фич)
# ffmpeg-python==0.2.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

